simplify producer
